Extract Op enum from Client
